Add Atribute

Adds an attribute element according to EML standards. Multiple attributes can be created. Will be appended to the attribute list.

add_attribute(
  attribute_name,
  attribute_definition,
  storage_type,
  measurement_scale,
  attribute_label = NULL,
  domain = NULL,
  definition = NULL,
  text_pattern = NULL,
  type = NULL,
  units = NULL,
  number_type = NULL,
  unit_precision = NULL,
  date_time_format = NULL,
  date_time_precision = NULL,
  minimum = NULL,
  maximum = NULL
)

Arguments

attribute_name
The name of a field in a data table. This is often a short and/or cryptic name. It is recommended that the attribute names be suitable for use as a variable, e.g., composed of ASCII characters, and that the attribute names match the column headers of a CSV or other text table.

storage_type
An indicator to processing systems as to how the attribute might be represented in a system or language, but is distinct from the actual expression of the domain of the attribute. A list of approved storage types can be found at storage_type

measurement_scale
The type of scale from which values are drawn for the attribute. A list of approved measurment scales can be viewed at measurement_scale. Further explination of the choices and required inputs are described below.

attribute_label
Optional. Used to provide a less ambiguous or less cryptic alternative identification than what is provided in attribute_name.

domain
Input for the non-numeric measurement scales only. Please list either "text" or "enumerated".

definition
Either the text definition or code definition. Both are appended using this parameter, but have different applications. Examples of both are listed below.

text_pattern
A regular expression pattern constraining the attribute.

type
Either "ratio" or "interval".

units
The units assigned to this attribute's values.

number_type
A list of possible options can be viewed at number_type.

unit_precision
How precise units are measured.

date_time_format
The format your date/time attribute is recorded in. ISO 8601 standard should be used (YYYY-MM-DD).

date_time_precision
To what level time is being measured.

minimum
Theoretical or allowable minimum value. Values can be larger than or equal to this number.

maximum
Theoretical or allowable maximum value. Values can be less than or equal to this number.

attribute_defenition
A precise and complete definition of the attribute being documented.

Value

The project or dataset list with an attribute list appended.

Measurement Scales

Different measurement scale values will indicate different inputs:

Non-numeric:
Please provide a domain to indicate if your attribute is text or enumerated. Nominal: Used to define categorical scale attributes. If your attribute falls under the domain of "text", please provide the inputs of definition and text_pattern. If your attribute falls under the domain of "enumerated", please provide the input of definition. Ordinal: Used to define ordered scale attributes. If your attribute falls under the domain of "text", please provide the inputs of definition and text_pattern. If your attribute falls under the domain of "enumerated", please provide the input of definition.
Numeric:
Interval: Used to define interval scale attributes. Please provide the inputs of type, units, unit_precision, number_type, minimum, and maximum. Ratio: Used to define ratio scale attributes. Please provide the inputs of type, units, unit_precision, number_type, minimum, and maximum. dateTime: Used to define date and time attributes. Please provide the inputs of date_time_format, date_time_precision, minimum, and maximum.
Examples

# Nominal(text):
add_attribute(attribute_name = "site_id",
              attribute_definition = "Site id as used in sites table",
              storage_type = cvpiaEDIutils::storage_type$integer,
              measurement_scale = cvpiaEDIutils::measurement_scale$nominal,
              domain= "text",
              definition = "Site id as used in sites table.")

#> No attribute label provided.

#> No text pattern is provided. Please add if applicable.

#> $attributeName
#> [1] "site_id"
#> 
#> $attributeDefinition
#> [1] "Site id as used in sites table"
#> 
#> $storageType
#> NULL
#> 
#> $measurementScale
#> $measurementScale$nominal
#> $measurementScale$nominal$nonNumericDomain
#> $measurementScale$nominal$nonNumericDomain$textDomain
#> $measurementScale$nominal$nonNumericDomain$textDomain$definition
#> [1] "Site id as used in sites table."
#> 
#> 
#> 
#> 
#> 

# Nominal(enumerated):
code_def_1 = list(code = "yes", definition = "has been captured previously")
code_def_2 = list(code = "no", definition = "has not been captured previously")
code_definition = list(code_def_1, code_def_2)
add_attribute(attribute_name = "Recap",
              attribute_definition = "Has the Turtle been captured and tagged previously",
              storage_type = cvpiaEDIutils::storage_type$text,
              measurement_scale = cvpiaEDIutils::measurement_scale$nominal,
              domain = "enumerated",
              definition = code_definition)

#> No attribute label provided.

#> $attributeName
#> [1] "Recap"
#> 
#> $attributeDefinition
#> [1] "Has the Turtle been captured and tagged previously"
#> 
#> $storageType
#> NULL
#> 
#> $measurementScale
#> $measurementScale$nominal
#> $measurementScale$nominal$nonNumericDomain
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain[[1]]
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain[[1]]$code
#> [1] "yes"
#> 
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain[[1]]$definition
#> [1] "has been captured previously"
#> 
#> 
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain[[2]]
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain[[2]]$code
#> [1] "no"
#> 
#> $measurementScale$nominal$nonNumericDomain$enumeratedDomain[[2]]$definition
#> [1] "has not been captured previously"
#> 
#> 
#> 
#> 
#> 
#> 

# Ordinal(text):
add_attribute(attribute_name = "LatitudeDD",
              attribute_definition = "Latitude",
              storage_type = cvpiaEDIutils::storage_type$string,
              measurement_scale = cvpiaEDIutils::measurement_scale$ordinal,
              domain= "text",
              definition = "Latitude")

#> No attribute label provided.

#> No text pattern is provided. Please add if applicable.

#> $attributeName
#> [1] "LatitudeDD"
#> 
#> $attributeDefinition
#> [1] "Latitude"
#> 
#> $storageType
#> [1] "string"
#> 
#> $measurementScale
#> $measurementScale$ordinal
#> $measurementScale$ordinal$nonNumericDomain
#> $measurementScale$ordinal$nonNumericDomain$textDomain
#> $measurementScale$ordinal$nonNumericDomain$textDomain$definition
#> [1] "Latitude"
#> 
#> 
#> 
#> 
#> 

# Ordinal(enumerated): 
code_def_0 = list(code = "0", definition = "0 insects per meter of branch")
code_def_1 = list(code = "1", definition = "1-10 insects per meter")
code_def_2 = list(code = "2", definition = "11 – 100 insects per meter")
code_def_3 = list(code = "3", definition = "more than 100 insects per meter")
code_definition = list(code_def_0, code_def_1, code_def_2, code_def_3)
add_attribute(attribute_name = "hwa",
              attribute_definition = "Hemlock woolly adelgid density per meter of branch",
              storage_type = cvpiaEDIutils::storage_type$decimal,
              measurement_scale = cvpiaEDIutils::measurement_scale$ordinal,
              domain = "enumerated",
              definition = code_definition)

#> No attribute label provided.

#> $attributeName
#> [1] "hwa"
#> 
#> $attributeDefinition
#> [1] "Hemlock woolly adelgid density per meter of branch"
#> 
#> $storageType
#> [1] "decimal"
#> 
#> $measurementScale
#> $measurementScale$ordinal
#> $measurementScale$ordinal$nonNumericDomain
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[1]]
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[1]]$code
#> [1] "0"
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[1]]$definition
#> [1] "0 insects per meter of branch"
#> 
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[2]]
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[2]]$code
#> [1] "1"
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[2]]$definition
#> [1] "1-10 insects per meter"
#> 
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[3]]
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[3]]$code
#> [1] "2"
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[3]]$definition
#> [1] "11 – 100 insects per meter"
#> 
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[4]]
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[4]]$code
#> [1] "3"
#> 
#> $measurementScale$ordinal$nonNumericDomain$enumeratedDomain[[4]]$definition
#> [1] "more than 100 insects per meter"
#> 
#> 
#> 
#> 
#> 
#> 

# Interval:
add_attribute(attribute_name = "Count",
              attribute_definition = "Number of individuals observed",
              measurement_scale = cvpiaEDIutils::measurement_scale$interval,
              storage_type = cvpiaEDIutils::storage_type$integer,
              type = "interval",
              units = "number",
              unit_precision = "1",
              number_type = cvpiaEDIutils::number_type$whole,
              minimum = "0")

#> No attribute label provided.

#> Warning: Please provide a maximum theoretical value if applicable.

#> $attributeName
#> [1] "Count"
#> 
#> $attributeDefinition
#> [1] "Number of individuals observed"
#> 
#> $storageType
#> NULL
#> 
#> $measurementScale
#> $measurementScale$interval
#> $measurementScale$interval$unit
#> $measurementScale$interval$unit$standardUnit
#> [1] "number"
#> 
#> 
#> $measurementScale$interval$precision
#> [1] "1"
#> 
#> $measurementScale$interval$numericDomain
#> $measurementScale$interval$numericDomain$numberType
#> [1] "whole"
#> 
#> $measurementScale$interval$numericDomain$bounds
#> $measurementScale$interval$numericDomain$bounds$minimum
#> $measurementScale$interval$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $measurementScale$interval$numericDomain$bounds$minimum$minimum
#> [1] "0"
#> 
#> 
#> 
#> 
#> 
#> 

# Ratio: 
add_attribute(attribute_name = "pH",
              attribute_definition = "pH of soil solution",
              storage_type = cvpiaEDIutils::storage_type$float,
              measurement_scale = cvpiaEDIutils::measurement_scale$ratio,
              type = "ratio",
              units = "dimensionless",
              unit_precision = "0.01",
              number_type = cvpiaEDIutils::number_type$real)

#> No attribute label provided.

#> Warning: Please provide a minimum theoretical value if applicable.

#> Warning: Please provide a minimum theoretical value if applicable.

#> $attributeName
#> [1] "pH"
#> 
#> $attributeDefinition
#> [1] "pH of soil solution"
#> 
#> $storageType
#> [1] "float"
#> 
#> $measurementScale
#> $measurementScale$ratio
#> $measurementScale$ratio$unit
#> $measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $measurementScale$ratio$precision
#> [1] "0.01"
#> 
#> $measurementScale$ratio$numericDomain
#> $measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> 
#> 
#> 

# dateTime:
add_attribute(attribute_name = "Yrs",
              attribute_definition = "Calendar year of the observation from years 1990 - 2010.",
              storage_type = cvpiaEDIutils::storage_type$integer,
              measurement_scale = cvpiaEDIutils::measurement_scale$dateTime,
              attribute_label = "Years",
              date_time_format = "YYYY",
              date_time_precision = "1",
              minimum = "1993",
              maximum = "2003")

#> $attributeName
#> [1] "Yrs"
#> 
#> $attributeDefinition
#> [1] "Calendar year of the observation from years 1990 - 2010."
#> 
#> $storageType
#> NULL
#> 
#> $attributeLabel
#> [1] "Years"
#> 
#> $measurementScale
#> $measurementScale$dateTime
#> $measurementScale$dateTime$formatString
#> [1] "YYYY"
#> 
#> $measurementScale$dateTime$dateTimePrecision
#> [1] "1"
#> 
#> $measurementScale$dateTime$dateTimeDomain
#> $measurementScale$dateTime$dateTimeDomain$bounds
#> $measurementScale$dateTime$dateTimeDomain$bounds$minimum
#> $measurementScale$dateTime$dateTimeDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $measurementScale$dateTime$dateTimeDomain$bounds$minimum$minimum
#> [1] "1993"
#> 
#> 
#> $measurementScale$dateTime$dateTimeDomain$bounds$maximum
#> $measurementScale$dateTime$dateTimeDomain$bounds$maximum$exclusive
#> [1] "false"
#> 
#> $measurementScale$dateTime$dateTimeDomain$bounds$maximum$maximum
#> [1] "2003"
#> 
#> 
#> 
#> 
#> 
#>