Add Coverage Element

Adds the coverage information of a dataset based off of EML standards.

add_coverage(
  parent_element,
  geographic_description,
  west_bounding_coordinate,
  east_bounding_coordinate,
  north_bounding_coordinate,
  south_bounding_coordinate,
  begin_date,
  end_date,
  taxonomic_coverage = NULL
)

Arguments

parent_element
A list representing the EML project or dataset.

geographic_description
A description of the locations of research sites and areas related to the data.

west_bounding_coordinate
The west cardinality limit.

east_bounding_coordinate
The east cardinality limit.

north_bounding_coordinate
The north cardinality limit.

south_bounding_coordinate
The south cardinality limit.

begin_date
The starting date for the dataset or project. Dates must be provided in ISO 8601 format, YYYY-MM-DD.

end_date
The projected or actual end date for the dataset or project. Dates must be provided in ISO 8601 format, YYYY-MM-DD.

taxonomic_coverage
Addition of taxonomic coverage is optional. The taxonomic information for all organisms relevant to the study should be included and mandates the input of the taxonomy levels of kingdom-species. An example of how to append this information using the add_taxonomic_coverage function is given down below.

Value

The dataset or project with coverage information appended.

Examples

add_coverage(parent_element = list(), geographic_description = "North Slope
             drainage basin: Bounding box encompasses 42 drainage basins
             totaling the North Slope drainage basin, Alaska, USA.",
             west_bounding_coordinate = "-160.594000",
             east_bounding_coordinate = "-134.104800",
             north_bounding_coordinate = "71.238300",
             south_bounding_coordinate = "67.865000",
             begin_date = "1980-01-01",
             end_date = "2010-12-31")

#> $coverage
#> $coverage$geographicCoverage
#> $coverage$geographicCoverage$geographicDescription
#> [1] "North Slope\n             drainage basin: Bounding box encompasses 42 drainage basins\n             totaling the North Slope drainage basin, Alaska, USA."
#> 
#> $coverage$geographicCoverage$boundingCoordinates
#> $coverage$geographicCoverage$boundingCoordinates$westBoundingCoordinate
#> [1] "-160.594000"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$eastBoundingCoordinate
#> [1] "-134.104800"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$northBoundingCoordinate
#> [1] "71.238300"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$southBoundingCoordinate
#> [1] "67.865000"
#> 
#> 
#> 
#> $coverage$temporalCoverage
#> $coverage$temporalCoverage$rangeOfDates
#> $coverage$temporalCoverage$rangeOfDates$beginDate
#> $coverage$temporalCoverage$rangeOfDates$beginDate$calendarDate
#> [1] "1980-01-01"
#> 
#> 
#> $coverage$temporalCoverage$rangeOfDates$endDate
#> $coverage$temporalCoverage$rangeOfDates$endDate$calendarDate
#> [1] "2010-12-31"
#> 
#> 
#> 
#> 
#> 

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = cvpiaEDIutils::CVPIA_common_species$chinook)
add_coverage(parent_element = list(), geographic_description = "Description",
             west_bounding_coordinate = "-160.594000",
             east_bounding_coordinate = "-134.104800",
             north_bounding_coordinate = "71.238300",
             south_bounding_coordinate = "67.865000",
             begin_date = "1980-01-01",
             end_date = "2010-12-31",
             taxonomic_coverage = taxonomic_coverage)

#> $coverage
#> $coverage$geographicCoverage
#> $coverage$geographicCoverage$geographicDescription
#> [1] "Description"
#> 
#> $coverage$geographicCoverage$boundingCoordinates
#> $coverage$geographicCoverage$boundingCoordinates$westBoundingCoordinate
#> [1] "-160.594000"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$eastBoundingCoordinate
#> [1] "-134.104800"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$northBoundingCoordinate
#> [1] "71.238300"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$southBoundingCoordinate
#> [1] "67.865000"
#> 
#> 
#> 
#> $coverage$temporalCoverage
#> $coverage$temporalCoverage$rangeOfDates
#> $coverage$temporalCoverage$rangeOfDates$beginDate
#> $coverage$temporalCoverage$rangeOfDates$beginDate$calendarDate
#> [1] "1980-01-01"
#> 
#> 
#> $coverage$temporalCoverage$rangeOfDates$endDate
#> $coverage$temporalCoverage$rangeOfDates$endDate$calendarDate
#> [1] "2010-12-31"
#> 
#> 
#> 
#> 
#> $coverage$taxonomicCoverage
#> $coverage$taxonomicCoverage$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonRankName
#> [1] "kingdom"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonRankValue
#> [1] "Animalia"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "phylum"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Chordata"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "class"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Teleostei"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "order"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Salmoniformes"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "family"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Salmonidae"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "genus"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Oncorhynchus"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "species"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Oncorhynchus tshawytscha"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$commonName
#> [1] "Chinook Salmon"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$provider
#> [1] "https://itis.gov"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$taxonId
#> [1] "161980"
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 

# For adding multiple taxonomies to the coverage: 
chinook <- add_taxonomic_coverage(CVPIA_common_species = cvpiaEDIutils::CVPIA_common_species$chinook)
delta <- add_taxonomic_coverage(CVPIA_common_species = cvpiaEDIutils::CVPIA_common_species$delta_smelt)
taxonomic_coverage <- list(chinook, delta)
add_coverage(parent_element = list(), geographic_description = "Description",
             west_bounding_coordinate = "-160.594000",
             east_bounding_coordinate = "-134.104800",
             north_bounding_coordinate = "71.238300",
             south_bounding_coordinate = "67.865000",
             begin_date = "1980-01-01",
             end_date = "2010-12-31",
             taxonomic_coverage = taxonomic_coverage)

#> $coverage
#> $coverage$geographicCoverage
#> $coverage$geographicCoverage$geographicDescription
#> [1] "Description"
#> 
#> $coverage$geographicCoverage$boundingCoordinates
#> $coverage$geographicCoverage$boundingCoordinates$westBoundingCoordinate
#> [1] "-160.594000"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$eastBoundingCoordinate
#> [1] "-134.104800"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$northBoundingCoordinate
#> [1] "71.238300"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$southBoundingCoordinate
#> [1] "67.865000"
#> 
#> 
#> 
#> $coverage$temporalCoverage
#> $coverage$temporalCoverage$rangeOfDates
#> $coverage$temporalCoverage$rangeOfDates$beginDate
#> $coverage$temporalCoverage$rangeOfDates$beginDate$calendarDate
#> [1] "1980-01-01"
#> 
#> 
#> $coverage$temporalCoverage$rangeOfDates$endDate
#> $coverage$temporalCoverage$rangeOfDates$endDate$calendarDate
#> [1] "2010-12-31"
#> 
#> 
#> 
#> 
#> $coverage$taxonomicCoverage
#> $coverage$taxonomicCoverage[[1]]
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonRankName
#> [1] "kingdom"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonRankValue
#> [1] "Animalia"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "phylum"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Chordata"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "class"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Teleostei"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "order"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Salmoniformes"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "family"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Salmonidae"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "genus"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Oncorhynchus"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "species"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Oncorhynchus tshawytscha"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$commonName
#> [1] "Chinook Salmon"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$provider
#> [1] "https://itis.gov"
#> 
#> $coverage$taxonomicCoverage[[1]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$taxonId
#> [1] "161980"
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> $coverage$taxonomicCoverage[[2]]
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonRankName
#> [1] "kingdom"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonRankValue
#> [1] "Animalia"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "phylum"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Chordata"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "class"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Teleostei"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "order"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Osmeriformes"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "family"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Osmeridae"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "genus"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Hypomesus"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "species"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Hypomesus transpacificus"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$commonName
#> [1] "Delta Smelt"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$provider
#> [1] "https://itis.gov"
#> 
#> $coverage$taxonomicCoverage[[2]]$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$taxonId
#> [1] "162032"
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#>