Add Funding Element

Adds the funding information of a dataset based off of EML standards.

add_funding(
  parent_element,
  funder_name,
  funder_identifier,
  award_number,
  award_title,
  award_url = NULL,
  funding_description = NULL
)

Arguments

parent_element
A list representing the EML project or dataset.

funder_name
Organization or individual providing the funding.

funder_identifier
This is where the funding organization is listed in the registry. The funder identifier must be registered. Follow the instructions at https://ror.org/curation/ to get registered.

award_number
The identifier assigned by the funding agency to identify this funding award.

award_title
Title of the dataset or project which recieved funding.

award_url
Optional to include a link to information about the funding award on the funding organization's webpage.

funding_description
Optional to provide a short description of the funding recieved.

Value

The dataset or project with funding information appended.

Examples

add_funding(parent_element = list(), funder_name = "National Science Foundation",
            funder_identifier = "http://dx.doi.org/10.13039/100000001",
            award_number = "1656026",
            award_title = "LTER: Beaufort Sea Lagoons: An Arctic Coastal Ecosystem in Transition",
            award_url = "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026",
            funding_description = "BLE LTER is supported by the National Science 
                                   Foundation under award #1656026 
                                   (2017-08-01 to 2022-07-31)." )

#> $funding
#> $funding$para
#> [1] "BLE LTER is supported by the National Science \n                                   Foundation under award #1656026 \n                                   (2017-08-01 to 2022-07-31)."
#> 
#> 
#> $award
#> $award$funderName
#> [1] "National Science Foundation"
#> 
#> $award$funderIdentifier
#> [1] "http://dx.doi.org/10.13039/100000001"
#> 
#> $award$awardNumber
#> [1] "1656026"
#> 
#> $award$title
#> [1] "LTER: Beaufort Sea Lagoons: An Arctic Coastal Ecosystem in Transition"
#> 
#> $award$awardUrl
#> [1] "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026"
#> 
#>