Add Interval or Ratio Measurement Scales

Add Interval or Ratio Measurement Scales

add_interval_ratio(
  type = c("interval", "ratio"),
  units,
  unit_precision,
  number_type,
  minimum = NULL,
  maximum = NULL
)

Arguments

type
Either "interval" or "ratio". Use "interval" to define data which consist of equidistant points on a scale. Use "ratio" to define data which consists not only of equidistant points but also has a meaningful zero point, which allows the ratio to have meaning.

units
The units assigned to this attribute.

unit_precision
How precise this attirbutes' measurements are recorded.

number_type
What type of number. Examples given in exported documentation.

minimum
Optional. Theoretical or allowable minimum value. Values can be larger than or equal to this number.

maximum
Optional. Theoretical or allowable maximum value. Values can be less than or equal to this number.

Examples

# Interval:
add_attribute(attribute_name = "Count",
              attribute_definition = "Number of individuals observed",
                   measurement_scale = EDIutils::measurement_scale$interval, 
                   storage_type = EDIutils::storage_type$integer,
                   type = "interval",
                   units = "number",
                   unit_precision = "1",
                   number_type = EDIutils::number_type$whole, 
                   minimum = "0")

#> No attribute label provided.

#> Warning: Please provide a maximum theoretical value if applicable.

#> $attributeName
#> [1] "Count"
#> 
#> $attributeDefinition
#> [1] "Number of individuals observed"
#> 
#> $storageType
#> NULL
#> 
#> $measurementScale
#> $measurementScale$interval
#> $measurementScale$interval$unit
#> $measurementScale$interval$unit$standardUnit
#> [1] "number"
#> 
#> 
#> $measurementScale$interval$precision
#> [1] "1"
#> 
#> $measurementScale$interval$numericDomain
#> $measurementScale$interval$numericDomain$numberType
#> [1] "whole"
#> 
#> $measurementScale$interval$numericDomain$bounds
#> $measurementScale$interval$numericDomain$bounds$minimum
#> $measurementScale$interval$numericDomain$bounds$minimum$exclusive
#> [1] "false"
#> 
#> $measurementScale$interval$numericDomain$bounds$minimum$minimum
#> [1] "0"
#> 
#> 
#> 
#> 
#> 
#> 

                   
# Ratio: 
add_attribute(attribute_name = "pH",
              attribute_definition = "pH of soil solution",
              storage_type = EDIutils::storage_type$float,
              measurement_scale = EDIutils::measurement_scale$ratio,
              type = "ratio",
              units = "dimensionless",
              unit_precision = "0.01",
              number_type = EDIutils::number_type$real)

#> No attribute label provided.

#> Warning: Please provide a minimum theoretical value if applicable.

#> Warning: Please provide a minimum theoretical value if applicable.

#> $attributeName
#> [1] "pH"
#> 
#> $attributeDefinition
#> [1] "pH of soil solution"
#> 
#> $storageType
#> [1] "float"
#> 
#> $measurementScale
#> $measurementScale$ratio
#> $measurementScale$ratio$unit
#> $measurementScale$ratio$unit$standardUnit
#> [1] "dimensionless"
#> 
#> 
#> $measurementScale$ratio$precision
#> [1] "0.01"
#> 
#> $measurementScale$ratio$numericDomain
#> $measurementScale$ratio$numericDomain$numberType
#> [1] "real"
#> 
#> 
#> 
#>