Add Personnel Element

Adds personel according to EML standards

add_personnel(
  parent_element,
  first_name,
  last_name,
  email,
  role,
  orcid = NULL,
  organization = NULL
)

Arguments

parent_element
A list representing the EML project or dataset

first_name
Person's given name

last_name
Person's surname

email
Person's email address

role
Use "Creator" if you are one of the primary originators of the data. Other possible roles "Data Manager", "Field Technician", or "Assistant Researcher". There can be multiple personnel on a project with the same role.

orcid
ORCID iD is a persistent digital identifier for researchers, register at http://orcid.org/

organization
Person's employer or the entity they are associated with for this dataset or project

Examples

add_personnel(parent_element = list(),
              first_name = 'Katherine',
              last_name = "Johnson",
              email = 'kjohnson@nasa.gov',
              role = 'Creator',
              organization = 'NASA',
              orcid = '12345')

#> $contact
#> $contact$individualName
#> $contact$individualName$givenName
#> [1] "Katherine"
#> 
#> $contact$individualName$surName
#> [1] "Johnson"
#> 
#> 
#> $contact$electronicMailAddress
#> [1] "kjohnson@nasa.gov"
#> 
#> $contact$organizationName
#> [1] "NASA"
#> 
#> 
#> $creator
#> $creator$individualName
#> $creator$individualName$givenName
#> [1] "Katherine"
#> 
#> $creator$individualName$surName
#> [1] "Johnson"
#> 
#> 
#> $creator$electronicMailAddress
#> [1] "kjohnson@nasa.gov"
#> 
#> $creator$organizationName
#> [1] "NASA"
#> 
#> $creator$`@id`
#> [1] "12345"
#> 
#> 

add_personnel(parent_element = list(),
              first_name = "Edith",
              last_name = "Windsor",
              email = 'ewindsor@ibm.com',
              role = 'Data Manager',
              organization = 'IBM')

#> $associatedParty
#> $associatedParty$individualName
#> $associatedParty$individualName$givenName
#> [1] "Edith"
#> 
#> $associatedParty$individualName$surName
#> [1] "Windsor"
#> 
#> 
#> $associatedParty$electronicMailAddress
#> [1] "ewindsor@ibm.com"
#> 
#> $associatedParty$organizationName
#> [1] "IBM"
#> 
#> $associatedParty$role
#> [1] "Data Manager"
#> 
#>