Add project element

This function creates a project node within the parent element that contains all the required elements for the project section of an EML document. This function can be used in combination with add_personnel and add_funding. add_personnel can be used to generate the project_personnel input and add_funding can be used to generate the award_information input.

add_project(
  parent_element,
  project_title,
  award_information,
  project_personnel
)

Arguments

parent_element
A list in which the project node should be nested in.

project_title
The title of the project that the funding is awarded to.

award_information
A list that includes the required funding information for an EML document.This list must include the award title and the funderName. This list can be created by calling the add_funding function on the funding information or by manually inputting the required information. If the list is written manually it must be formatted as follows.

award_infomation = list(funderName = "Name", title = "Award Title")

Additional information about the funding may be added to the list. See the add_funding documentation for more information.

project_personnel
A list that includes the required information on project personnel for an EML document. It must include the first name, last name, organization, and personnel role for this project. This list can be created by calling the add_personnel function on the project personnel or by manually inputting the required information. If the list is written manually it must be formatted as follows.

project_personnel = list(individualName = list(givenName = "First Name", surName = "Last Name"), role = "Position", organization = "Organization")

Additional information about the project personnel may be added to the list. See the add_personnel documentation for more information.

Value

This function returns the parent element with a new project node containing all project information required for an EML document.

Examples

add_project(parent_element = list(), 
            project_title = "my project title", 
            award_information = add_funding(funder_name = "Bank",
                                            funder_identifier = "Funder 1",
                                            award_number = "000",
                                            award_title = "Money up for grabs",
                                            award_url = "awardforme.com"),
            project_personnel = add_personnel(parent_element = list(),
                                              first_name = "Smithy",
                                              last_name = "Smith",
                                              email = "myemail@mail.gov",
                                              role = "Manager",
                                              organization = "US GOV"))

#> Warning: Please provide the funding_description

#> $project
#> $project$title
#> [1] "my project title"
#> 
#> $project$personnel
#> $project$personnel$associatedParty
#> $project$personnel$associatedParty$individualName
#> $project$personnel$associatedParty$individualName$givenName
#> [1] "Smithy"
#> 
#> $project$personnel$associatedParty$individualName$surName
#> [1] "Smith"
#> 
#> 
#> $project$personnel$associatedParty$electronicMailAddress
#> [1] "myemail@mail.gov"
#> 
#> $project$personnel$associatedParty$organizationName
#> [1] "US GOV"
#> 
#> $project$personnel$associatedParty$role
#> [1] "Manager"
#> 
#> 
#> 
#> $project$award
#> $project$award$funderName
#> [1] "Bank"
#> 
#> $project$award$title
#> [1] "Money up for grabs"
#> 
#> $project$award$funderIdentifier
#> [1] "Funder 1"
#> 
#> $project$award$awardNumber
#> [1] "000"
#> 
#> $project$award$awardUrl
#> [1] "awardforme.com"
#> 
#> 
#> 

                                              
add_project(parent_element = list(), 
            project_title = "my project title", 
            award_information = list(funderName = "Bank",
                                     funderIdentifier = "Funder 1",
                                     awardNumber = "000",
                                     title = "Money up for grabs",
                                     awardUrl = "awardforme.com"),
            project_personnel = list(individualName = list(givenName = "Smithy",
                                                           surName = "Smith"),
                                     electronicMailAddress = "myemail@mail.gov",
                                     role = "Manager",
                                     organizationName = "US GOV"))

#> $project
#> $project$title
#> [1] "my project title"
#> 
#> $project$personnel
#> $project$personnel$individualName
#> $project$personnel$individualName$givenName
#> [1] "Smithy"
#> 
#> $project$personnel$individualName$surName
#> [1] "Smith"
#> 
#> 
#> $project$personnel$electronicMailAddress
#> [1] "myemail@mail.gov"
#> 
#> $project$personnel$role
#> [1] "Manager"
#> 
#> $project$personnel$organizationName
#> [1] "US GOV"
#> 
#> 
#> $project$award
#> $project$award$funderName
#> [1] "Bank"
#> 
#> $project$award$funderIdentifier
#> [1] "Funder 1"
#> 
#> $project$award$awardNumber
#> [1] "000"
#> 
#> $project$award$title
#> [1] "Money up for grabs"
#> 
#> $project$award$awardUrl
#> [1] "awardforme.com"
#> 
#> 
#>