Get recent uploads

Get recent uploads

get_recent_uploads(query, env = "production")

Arguments

query
(character) Query (see details below)

env
(character) Repository environment. Can be: "production", "staging", or "development".

Value

(xml_document) A list of zero or more audit records of either recently inserted or recently updated data packages.

Details

Query parameters are specified as key=value pairs, multiple pairs must be delimited with ampersands (&), and only a single value should be specified for a particular key. The following query parameter keys are allowed:

serviceMethod - Can be: createDataPackage, updateDataPackage

fromTime - An ISO8601 timestamp

limit - A positive whole number

The query parameter serviceMethod should have the value "createDataPackage" (to retrieve recent inserts) or "updateDataPackage" (to retrieve recent updates). The query parameter fromTime is used to specify the date/time in the past that represents the oldest audit records that should be returned. Data packages uploaded prior to that time are not considered recent uploads and are thus filtered from the query results. The query parameter limit sets an upper limit on the number of audit records returned. For example, "limit=3".

Examples

# Get the 5 most recently created data packages
auditReport <- get_recent_uploads(
  query = "serviceMethod=createDataPackage&limit=5")

#> Error in get_recent_uploads(query = "serviceMethod=createDataPackage&limit=5"): could not find function "get_recent_uploads"

# Show the first
xml2::xml_find_first(auditReport, "auditRecord")

#> Error in xml2::xml_find_first(auditReport, "auditRecord"): object 'auditReport' not found