List active reservations

List active reservations

list_active_reservations(env = "production")

Arguments

env

(character) Repository environment. Can be: "production", "staging", or "development".

Value

(xml_document) The set of data package identifiers that users have actively reserved. Note that data package identifiers that have been successfully uploaded are no longer considered active reservations and thus are not included in this list.

See also

Examples

# List reservations
reservations <- list_active_reservations()
reservations
#> {xml_document}
#> <reservations>
#>  [1] <reservation>\n  <docid>edi.11</docid>\n  <principal>uid=NTL,o=LTER,dc=e ...
#>  [2] <reservation>\n  <docid>edi.130</docid>\n  <principal>uid=user23,o=EDI,d ...
#>  [3] <reservation>\n  <docid>edi.131</docid>\n  <principal>uid=user20,o=EDI,d ...
#>  [4] <reservation>\n  <docid>edi.132</docid>\n  <principal>uid=chase,o=EDI,dc ...
#>  [5] <reservation>\n  <docid>edi.142</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#>  [6] <reservation>\n  <docid>edi.152</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#>  [7] <reservation>\n  <docid>edi.154</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#>  [8] <reservation>\n  <docid>edi.156</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#>  [9] <reservation>\n  <docid>edi.158</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [10] <reservation>\n  <docid>edi.159</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [11] <reservation>\n  <docid>edi.161</docid>\n  <principal>uid=jtallant,o=EDI ...
#> [12] <reservation>\n  <docid>edi.162</docid>\n  <principal>uid=jtallant,o=EDI ...
#> [13] <reservation>\n  <docid>edi.171</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [14] <reservation>\n  <docid>edi.172</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [15] <reservation>\n  <docid>edi.173</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [16] <reservation>\n  <docid>edi.174</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [17] <reservation>\n  <docid>edi.175</docid>\n  <principal>uid=ahall,o=EDI,dc ...
#> [18] <reservation>\n  <docid>edi.177</docid>\n  <principal>uid=jtallant,o=EDI ...
#> [19] <reservation>\n  <docid>edi.178</docid>\n  <principal>uid=EDI,o=EDI,dc=e ...
#> [20] <reservation>\n  <docid>edi.180</docid>\n  <principal>uid=EDI,o=EDI,dc=e ...
#> ...

# Show first
xml2::xml_find_first(reservations, "reservation")
#> {xml_node}
#> <reservation>
#> [1] <docid>edi.11</docid>
#> [2] <principal>uid=NTL,o=LTER,dc=ecoinformatics,dc=org</principal>
#> [3] <dateReserved>2017-03-02 15:02:28.899</dateReserved>