List reservation identifiers

List reservation identifiers

list_reservation_identifiers(scope, env = "production")

Arguments

scope
(character) Scope of data package

env
(character) Repository environment. Can be: "production", "staging", or "development".

Value

(numeric) The set of identifiers for the specified scope that end users have actively reserved for future upload

See also

Examples

# List reservations
reservations <- list_reservation_identifiers(scope = "edi")
reservations

#>   [1]   11  130  131  132  142  152  154  156  158  159  161  162  171  172  173
#>  [16]  174  175  177  178  180  182  183  185  196  203  204  205  209  211  212
#>  [31]  245  249  266  288  293  294  297  298  300  301  304  310  315  360  429
#>  [46]  431  432  433  434  446  449  483  485  501  518  526  528  550  564  578
#>  [61]  579  586  588  589  604  630  636  639  655  656  657  658  659  660  661
#>  [76]  662  663  664  665  666  667  668  669  670  671  672  673  674  675  676
#>  [91]  678  680  681  683  684  685  686  687  689  690  693  697  717  718  719
#> [106]  720  721  727  728  729  741  750  757  758  761  762  763  764  765  766
#> [121]  767  768  769  770  771  772  773  774  775  776  777  778  779  780  781
#> [136]  782  783  784  785  786  787  788  789  790  791  792  793  794  795  796
#> [151]  797  798  799  800  801  802  803  804  805  806  807  808  809  810  811
#> [166]  812  813  814  815  816  819  823  830  835  844  847  856  883  891  899
#> [181]  957  961  977  990 1002 1014 1021 1040 1042 1048 1060 1061