Read data entity

Read data entity

read_data_entity(packageId, entityId, env = "production")

Arguments

packageId
(character) Data package identifier

entityId
(character) Data entity identifier

env
(character) Repository environment. Can be: "production", "staging", or "development".

Value

(raw) Raw bytes (i.e. application/octet-stream) to be parsed by a reader function appropriate for the data type

See also

Examples

# Read names and IDs of data entities in package "edi.1047.1"
res <- read_data_entity_names(packageId = "edi.1047.1")
res

#>                           entityId                entityName
#> 1 3abac5f99ecc1585879178a355176f6d        Environmentals.csv
#> 2 f6bfa89b48ced8292840e53567cbf0c8               ByCatch.csv
#> 3 c75642ddccb4301327b4b1a86bdee906               Chinook.csv
#> 4 2c9ee86cc3f3ffc729c5f18bfe0a2a1d             Steelhead.csv
#> 5 785690848dd20f4910637250cdc96819 TrapEfficiencyRelease.csv
#> 6 58b9000439a5671ea7fe13212e889ba5 TrapEfficiencySummary.csv
#> 7 86e61c1a501b7dcf0040d10e009bfd87        TrapOperations.csv


# Read raw bytes of the 3rd data entity
raw <- read_data_entity(packageId = "edi.1047.1", entityId = res$entityId[3])
head(raw)

#> [1] ef bb bf 44 61 74


# Parse with .csv reader
data <- readr::read_csv(file = raw)

#> indexed 3.00B in  0s, ?indexed 2.15GB in  0s, 2.15GB/s                                                                              

#> Warning: One or more parsing issues, see `problems()` for details

#> Rows: 105325 Columns: 20

#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr (12): Date, subSiteName, commonName, atCaptureRun, finalRun, finalRunMet...
#> dbl  (7): trapVisitID, catchRawID, releaseID, n, forkLength, weight, specime...
#> lgl  (1): Specimen.comments

#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.

data

#> # A tibble: 105,325 x 20
#>    Date          trapVisitID subSiteName   catchRawID releaseID commonName     n
#>    <chr>               <dbl> <chr>              <dbl>     <dbl> <chr>      <dbl>
#>  1 1/8/2014 0:00         330 North Channe~      32409         0 Chinook s~     1
#>  2 1/8/2014 0:00         330 North Channe~      32412         0 Chinook s~     1
#>  3 1/8/2014 0:00         330 North Channe~      32410         0 Chinook s~     1
#>  4 1/8/2014 0:00         330 North Channe~      32408         0 Chinook s~     1
#>  5 1/8/2014 0:00         330 North Channe~      32406         0 Chinook s~     1
#>  6 1/8/2014 0:00         322 North Channe~      31958         0 Chinook s~     1
#>  7 1/8/2014 0:00         322 North Channe~      31975         0 Chinook s~     1
#>  8 1/8/2014 0:00         322 North Channe~      31974         0 Chinook s~     1
#>  9 1/8/2014 0:00         322 North Channe~      31973         0 Chinook s~     1
#> 10 1/8/2014 0:00         322 North Channe~      31972         0 Chinook s~     1
#> # ... with 105,315 more rows, and 13 more variables: atCaptureRun <chr>,
#> #   finalRun <chr>, finalRunMethod <chr>, lifeStage <chr>, forkLength <dbl>,
#> #   weight <dbl>, mort <chr>, fishOrigin <chr>, markType <chr>,
#> #   CatchRaw.comments <chr>, specimenTypeID <dbl>, physicalSpecimenCode <chr>,
#> #   Specimen.comments <lgl>