Replace missing value codes with a new value

Missing value codes vary across data packages and some computational environments have preferences (e.g. R uses NA). This function transform these codes to make downstream use simpler.

replace_missing_value_codes(
  x = NULL,
  eml = NULL,
  with = NA,
  file = NULL,
  coerce = TRUE
)

Arguments

x
(data frame) Data with missing value codes to convert to NA.

eml
(xml_document xml_node) EML metadata listing missing value codes for x. Use EDIutils::api_read_metadata() or xml2::read_xml() to read the EML file.

with
(character or NA) New missing value code. This value will replace those listed in x as specified by eml. Default is NA.

file
(character) Name of the file, which x was created from (e.g. "Resin_Deposition.csv". Use this argument if a column in x has more than one match in x (see details).

coerce
(logical) Columns of x defined as numeric in eml will be coerced to numeric after missing value code replacement. Default is TRUE.

Value

(data frame) x with missing value codes replaced with with.

Details

For each column in x a matching attribute is searched for in eml. If one is found, then the missing value code listed in eml is replaced in the data with with. If more than one attribute is found then an error is returned prompting the use of the file argument. file maps to /eml/dataset/dataTable/physical/objectName in eml and creates an unambiguous column name/attribute match.

Columns in x that are defined in eml as numeric, but are read in as character due to the missing value code, are converted to numeric.