Update data package

Update data package

update_data_package(eml, useChecksum = FALSE, env = "production")

Arguments

eml
(character) Full path to an EML file describing the data package to be updated

useChecksum
(logical) Use data entities from a previous version of the data package? See details below.

env
(character) Repository environment. Can be: "production", "staging", or "development".

Value

transaction (character) Transaction identifier. May be used in a subsequent call to check_status_update() to determine the operation status

Details

Each data entity described in eml must be accompanied by a web accessible URL at the XPath ".//physical/distribution/online/url". The EDI data repository uses these links to download the data entities. The URLs must be static and not have any redirects otherwise the data entities will not be downloadable.

Note

User authentication is required (see login())

See also

Examples

if (FALSE) {

login()

# Update data package
transaction <- update_data_package(
  eml = "./data/edi.595.2.xml",
  env = "staging"
)
transaction
#> [1] "update_edi.595_163966788658131920__edi.595.2"

# Check update status
status <- check_status_update(
  transaction = transaction,
  env = "staging"
)
status
#> [1] TRUE

logout()
}