Add Funding Element

Creates the award information of a project based off of EML standards. This award element is then nested within a project node to complete a funding section.

add_funding(
  funder_name,
  award_title,
  funder_identifier = NULL,
  award_number = NULL,
  award_url = NULL,
  funding_description = NULL
)

Arguments

funder_name
Organization or individual providing the funding. If applicable, use one of the following from the helper data CVPIA_funders: "USBR", "CDWR", or "CDFW" to get pre-selected funding information. If you are funded by one of these sources please specify which one in funder_name and include an award_title.

award_title
Title of the award granted to the project.

funder_identifier
(Optional) This is where the funding organization is listed in the registry. The funder identifier must be registered. Follow the instructions at https://ror.org/curation/ to get registered.

award_number
(Optional) The identifier assigned by the funding agency to identify this funding award.

award_url
(Optional) Optional to include a link to information about the funding award on the funding organization's webpage.

funding_description
(Optional) Provide a short description of the funding received.

Value

An award list that is then added to the project element of an EML file.

Examples

add_funding(funder_name = "National Science Foundation",
            funder_identifier = "http://dx.doi.org/10.13039/100000001",
            award_number = "1656026",
            award_title = "National Science Foundation Grant",
            award_url = "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026",
            funding_description = "BLE LTER is supported by the National Science 
                                   Foundation under award #1656026 
                                   (2017-08-01 to 2022-07-31)." )

#> $funderName
#> [1] "National Science Foundation"
#> 
#> $title
#> [1] "National Science Foundation Grant"
#> 
#> $funderIdentifier
#> [1] "http://dx.doi.org/10.13039/100000001"
#> 
#> $awardNumber
#> [1] "1656026"
#> 
#> $description
#> [1] "BLE LTER is supported by the National Science \n                                   Foundation under award #1656026 \n                                   (2017-08-01 to 2022-07-31)."
#> 
#> $awardUrl
#> [1] "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026"
#>