Add Taxonomic Coverage

Adds the taxonomic coverage information of a dataset based off of EML standards. The addition of taxonomic coverage is optional, however defaults of chinook, sturgeon, smelt, and steelhead are provided. While single or multiple taxonomies can be applied, the full set of information must be provided if chosen to be included.

add_taxonomic_coverage(
  CVPIA_common_species = NULL,
  kingdom = "kingdom",
  kingdom_value,
  phylum = "phylum",
  phylum_value,
  class = "class",
  class_value,
  order = "order",
  order_value,
  family = "family",
  family_value,
  genus = "genus",
  genus_value,
  species = "species",
  species_value,
  common_name,
  taxon_id = NULL
)

Arguments

CVPIA_common_species
Use one of the following from the helper data CVPIA_common_species: "chinook", "delta_smelt", "white_sturgeon", "green_sturgeon", or "steelhead" to get pre-selected information from ITIS.

kingdom
Kingdom level present.

kingdom_value
The kingdom name.

phylum
Phylum level present.

phylum_value
The phylum name.

class
Class level present.

class_value
The class level name.

order
Order level present.

order_value
The order level name.

family
Family level present.

family_value
The family level name.

genus
Genus level present.

genus_value
The genus level name.

species
Species level present.

species_value
The species level name.

common_name
The common name of the organism.

taxon_id
Optional. The taxonomic serial number provided by ITIS.

Value

Taxonomic coverage information. The function should be assigned to the name taxonomic_coverage to append it to the dataset or project. Example of how to incorporate it into the add_coverage function is seen below.

CVPIA Common Species

The following frequently cited species are available for convenience:

EDIutils::CVPIA_common_species$chinook - Oncorhynchus tshawytscha (ITIS: 161980)

EDIutils::CVPIA_common_species$steelhead - Oncorhynchus mykiss (ITIS: 161989)

EDIutils::CVPIA_common_species$delta_smelt - Hypomesus transpacificus (ITIS: 162032)

EDIutils::CVPIA_common_species$white_sturgeon - Acipenser transmontanus (ITIS:161068)

EDIutils::CVPIA_common_species$green_sturgeon - Acipenser medirostris (ITIS: 161067)

For further taxonomic coverage (i.e subkingdom, infrakingdom, etc.) on any of these species, you can visit: ITIS's webpage for full coverage information.

Examples

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "chinook")

#> No taxon id has been provided. This number can be found at ITIS.gov if you wish to append it.

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "delta_smelt")

#> No taxon id has been provided. This number can be found at ITIS.gov if you wish to append it.

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "white_sturgeon")

#> No taxon id has been provided. This number can be found at ITIS.gov if you wish to append it.

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "green_sturgeon")

#> No taxon id has been provided. This number can be found at ITIS.gov if you wish to append it.

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "steelhead")

#> No taxon id has been provided. This number can be found at ITIS.gov if you wish to append it.

taxonomic_coverage <- add_taxonomic_coverage(kingdom_value = "Animalia",
                       phylum_value = "Chordata",
                       class_value = "Mammalia",
                       order_value = "Carnivora",
                       family_value = "Felidae",
                       genus_value = "Panthera", 
                       species_value = "Panthera Leo",
                       common_name = "Lion", 
                       taxon_id = "183803")   
                                            
# To append this information to the dataset or project:                        
    add_coverage(parent_element = list(), geographic_description = "Description",
                 west_bounding_coordinate = "-160.594000", 
                 east_bounding_coordinate = "-134.104800",
                 north_bounding_coordinate = "71.238300", 
                 south_bounding_coordinate = "67.865000",
                 begin_date = "1980-01-01", 
                 end_date = "2010-12-31", 
                 taxonomic_coverage = taxonomic_coverage)

#> $coverage
#> $coverage$geographicCoverage
#> $coverage$geographicCoverage$geographicDescription
#> [1] "Description"
#> 
#> $coverage$geographicCoverage$boundingCoordinates
#> $coverage$geographicCoverage$boundingCoordinates$westBoundingCoordinate
#> [1] "-160.594000"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$eastBoundingCoordinate
#> [1] "-134.104800"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$northBoundingCoordinate
#> [1] "71.238300"
#> 
#> $coverage$geographicCoverage$boundingCoordinates$southBoundingCoordinate
#> [1] "67.865000"
#> 
#> 
#> 
#> $coverage$temporalCoverage
#> $coverage$temporalCoverage$rangeOfDates
#> $coverage$temporalCoverage$rangeOfDates$beginDate
#> $coverage$temporalCoverage$rangeOfDates$beginDate$calendarDate
#> [1] "1980-01-01"
#> 
#> 
#> $coverage$temporalCoverage$rangeOfDates$endDate
#> $coverage$temporalCoverage$rangeOfDates$endDate$calendarDate
#> [1] "2010-12-31"
#> 
#> 
#> 
#> 
#> $coverage$taxonomicCoverage
#> $coverage$taxonomicCoverage$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonRankName
#> [1] "kingdom"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonRankValue
#> [1] "Animalia"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "phylum"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Chordata"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "class"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Mammalia"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "order"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Carnivora"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "family"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Felidae"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "genus"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Panthera"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
#> [1] "species"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
#> [1] "Panthera Leo"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$commonName
#> [1] "Lion"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$provider
#> [1] "https://itis.gov"
#> 
#> $coverage$taxonomicCoverage$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$taxonId
#> [1] "183803"
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#>