Get audit report

Get audit report

Usage

get_audit_report(query, as = "data.frame", env = "production")

Arguments

query

(character) Query (see details below)

as

(character) Format of the returned object. Can be: "data.frame" or "xml".

env

(character) Repository environment. Can be: "production", "staging", or "development".

Value

(data.frame or xml_document) Zero or more audit records matching the query parameters as specified in the request (see details below).

Details

Query parameters are specified as key=value pairs, multiple pairs must be delimited with ampersands (&), and only a single value should be specified for a particular key. The following query parameter keys are allowed:

category - Can be: debug, info, error, warn

service - Any of the EDI data repository services

serviceMethod - Any of the EDI data repository service Resource class JAX-RS methods

user - Any user

group - Any group

authSystem - A valid auth system identifier

status - A valid HTTP Response Code

resourceId - An EDI data repository resource identifier, e.g. https://pasta.lternet.edu/package/eml/knb-lter-and/2719/6, or a thereof (see details below)

fromTime - An ISO8601 timestamp

toTime - An ISO8601 timestamp

limit - A positive whole number

The query parameters fromTime and optionally toTime should be used to indicate a time span. When toTime is absent, the report will consist of all matching records up to the current time. Either of these parameters may only be used once. The query parameter limit sets an upper limit on the number of audit records returned. For example, "limit=1000". The query parameter resourceId will match any audit log entry whose resourceId value contains the specified string value. Thus, a query parameter of "resourceId=knb-lter-and" will match any audit log entry whose resourceId value contains the substring "knb-lter-and", while a query parameter of "resourceId=knb-lter-and/2719/6" will match any audit log entry whose resourceId value contains the substring "knb-lter-and/2719/6".

Note

User authentication is required (see login())

See also

Examples

if (FALSE) {

login()

# Get audit report for data reads between 2021-12-01 and 2021-12-02
query <- "serviceMethod=readDataEntity&fromTime=2021-12-01&toTime=2021-12-02"
auditReport <- get_audit_report(query)

logout()
}