List user data packages

List all data packages (including their revision values) uploaded to the repository by a particular user, specified by a distinguished name. Data packages that were uploaded by the specified user but have since been deleted are excluded from the list.

list_user_data_packages(dn, env = "production")

Arguments

dn

(character) Distinguished name of user. Create with create_dn().

env

(character) Repository environment. Can be: "production", "staging", or "development".

Value

(character) Data package identifiers belonging to a dn

See also

Examples

if (FALSE) {

# List user data packages
dn <- create_dn(userId = "dbjourneynorth")
packageIds <- list_user_data_packages(dn)
packageIds
#> [1] "edi.948.1" "edi.949.1"
}