Search data packages

Searches data packages in the EDI data repository using the specified Solr query.

search_data_packages(query, env = "production")

Arguments

query

(character) Query (see details below)

env

(character) Repository environment. Can be: "production", "staging", or "development".

Value

(xml_document) Search results containing the fields:

abstract

begindate

doi

enddate

funding

geographicdescription

id

methods

packageid

pubdate

responsibleParties

scope

site

taxonomic

title

authors

spatialCoverage

sources

keywords

organizations

singledates

timescales

Details

Documents in the EDI data repository Solr index can be discovered based on metadata values stored in the following list of searchable fields (not all EML content is queryable):

Single-value fields:

abstract

begindate

doi

enddate

funding

geographicdescription

id

methods

packageid

pubdate

responsibleParties

scope

singledate

site

taxonomic

title

Multi-value fields:

author

coordinates

keyword

organization

projectTitle

relatedProjectTitle

timescale

query parser: The optimal query parser (defType=edismax) is added to every query.

See Apache Solr Wiki for how to construct a Solr query.

Note

Only the newest version of data packages are searchable, older versions are not.

When constructing a query note that the 15403 data packages of the ecotrends project and 10492 data packages of the LTER Landsat project, can be excluded from the returned results by including &fq=-scope:(ecotrends+lter-landsat) in the query string.

Examples

# Search for data packages containing the term "air temperature"
search_data_packages(query = 'q="air+temperature"&fl=*')
#> {xml_document}
#> <resultset numFound="1344" start="0" rows="10">
#>  [1] <document>\n  <abstract>This project contains data collected from sites  ...
#>  [2] <document>\n  <abstract>This data set contains measurements of air tempe ...
#>  [3] <document>\n  <abstract>Air temperature is measured at 15-minute interva ...
#>  [4] <document>\n  <abstract>This data package contains 5-minute air temperat ...
#>  [5] <document>\n  <abstract>This data package contains 5-minute air temperat ...
#>  [6] <document>\n  <abstract>This data package contains 5-minute air temperat ...
#>  [7] <document>\n  <abstract>This data package contains 5-minute air temperat ...
#>  [8] <document>\n  <abstract>This data package contains 5-minute air temperat ...
#>  [9] <document>\n  <abstract>This data package contains 5-minute air temperat ...
#> [10] <document>\n  <abstract>This data package contains 5-minute air temperat ...

# Search for data packages containing the term "air temperature" and
# returning only the packageid, title, and score of each match
search_data_packages(query = 'q="air+temperature"&fl=packageid,title,score')
#> {xml_document}
#> <resultset numFound="1344" start="0" rows="10">
#>  [1] <document>\n  <packageid>knb-lter-cwt.4038.19</packageid>\n  <title>Weat ...
#>  [2] <document>\n  <packageid>knb-lter-cwt.4100.19</packageid>\n  <title>Weat ...
#>  [3] <document>\n  <packageid>knb-lter-hbr.322.1</packageid>\n  <title>Hubbar ...
#>  [4] <document>\n  <packageid>knb-lter-jrn.210548010.20</packageid>\n  <title ...
#>  [5] <document>\n  <packageid>knb-lter-jrn.210548011.22</packageid>\n  <title ...
#>  [6] <document>\n  <packageid>knb-lter-jrn.210548014.20</packageid>\n  <title ...
#>  [7] <document>\n  <packageid>knb-lter-jrn.210548015.20</packageid>\n  <title ...
#>  [8] <document>\n  <packageid>knb-lter-jrn.210548003.20</packageid>\n  <title ...
#>  [9] <document>\n  <packageid>knb-lter-jrn.210548002.20</packageid>\n  <title ...
#> [10] <document>\n  <packageid>knb-lter-jrn.210548001.21</packageid>\n  <title ...

# Search for data packages containing the term "air temperature", returning
# only the packageid, title, score, and excluding ecotrends and lter-landsat
# scopes from the returned results
query <- 'q="air+temperature"&fl=packageid,title,score&fq=-scope:(ecotrends+lter-landsat)'
search_data_packages(query)
#> {xml_document}
#> <resultset numFound="914" start="0" rows="10">
#>  [1] <document>\n  <packageid>knb-lter-cwt.4038.19</packageid>\n  <title>Weat ...
#>  [2] <document>\n  <packageid>knb-lter-cwt.4100.19</packageid>\n  <title>Weat ...
#>  [3] <document>\n  <packageid>knb-lter-hbr.322.1</packageid>\n  <title>Hubbar ...
#>  [4] <document>\n  <packageid>knb-lter-jrn.210548010.20</packageid>\n  <title ...
#>  [5] <document>\n  <packageid>knb-lter-jrn.210548011.22</packageid>\n  <title ...
#>  [6] <document>\n  <packageid>knb-lter-jrn.210548014.20</packageid>\n  <title ...
#>  [7] <document>\n  <packageid>knb-lter-jrn.210548015.20</packageid>\n  <title ...
#>  [8] <document>\n  <packageid>knb-lter-jrn.210548003.20</packageid>\n  <title ...
#>  [9] <document>\n  <packageid>knb-lter-jrn.210548002.20</packageid>\n  <title ...
#> [10] <document>\n  <packageid>knb-lter-jrn.210548001.21</packageid>\n  <title ...