Attribute Storage Types

A set of common options for different attribute storage types. A helper data object to be used with add_attribute

storage_type

Format

A named list, keys are named the same as values.

Character Storage Types

string: The string datatype represents character strings in XML. Use this link for more information.

Locical Storage Types

boolean: The boolean datatype has the value space required to support the mathematical concept of binary-valued logic i.e true, false. Use this link for more information.

Numeric Storage Types

decimal: The decimal datatype represents a subset of the real numbers, which can be represented by decimal numerals. Out of the 3 numeric storage types, decimal is the most precise. It uses up to 28-29 significant digits. Use this link for more information. Examples include: 2.00, -1.23, 0.0002, 210., 210.0

float: The float datatype is patterned after the IEEE single-precision 32-bit floating point type. It can be further represented as either a whole plus fractional digits (like decimal values) or as a mantissa plus an exponent. It uses up to 7 significant digits. Use this link for more information. Examples include decimals and/or 1.23E4, 1.2e-39

double: The double datatype is patterned after the IEEE double-precision 64-bit floating point type. The level of precision is the main difference between decimal, float, and double. It uses up to 15-16 significant digits. Use this link for more information. Examples include 3.14, 2.718

Time Storage Types

duration:The duration datatype represents a duration of time. The value space of duration is a six-dimensional space where the coordinates designate the Gregorian year, month, day, hour, minute, and second components defined in ISO 8601. Use this link for more information. For example, to indicate a duration of 1 year, 2 months, 3 days, 10 hours, and 30 minutes, one would write: P1Y2M3DT10H30M. One could also indicate a duration of minus 120 days as: -P120D.

dateTime:The dateTime datatype values may be viewed as objects with integer-valued year, month, day, hour and minute properties, a decimal-valued second property, and a boolean timezoned property. Use this link for more information. The typical format is: YYYY '-' MM '-' DD 'T' hh ':' mm ':' ss. An example of this format is: 2002-10-10T12:00:00-05:00 .

time: The time datatype represents an instant of time that recurs every day. Use this link for more information. The time datatype follows the format hh:mm:ss.sss, i.e 06:20:00.098

date: The date datatype consists of top-open intervals of exactly one day in length on the timelines of dateTime, beginning on the beginning moment of each day (in each timezone), i.e. '00:00:00', up to but not including '24:00:00'. Use this link for more information. The date datatype follows the format YYYY '-' MM '-' DD, i.e 2012-02-02.

gYearMonth: The gYearMonth datatype represents a specific gregorian month in a specific gregorian year. Use this link for more information. The datatype gYearMonth follows the format of CCYY-MM. For example May 1999 would be inputted as 1999-05.

gYear: The gYear datatype represents a gregorian calendar year. Use this link for more information. The datatype gYear follows the format of CCYY. For example the year 1999 would be inputted as 1999.

gMonthDay: The gMonthDay datatype is a gregorian date that recurs, specifically a day of the year such as the third of May. Use this link for more information. The gMonthDay datatype follows the format of MM-DD. For example, March 3rd would be inputted as 03-03.

gDay: The gDay datatype is a gregorian day that recurs, specifically a day of the month such as the 5th of the month. Use this link for more information. The gDay datatype follows the format -DD. For example, the 4th of a month would be inputted as 04.

gMonth: The gMonth datatype is a gregorian month that recurs every year. Use this link for more information. The gMonth datatype follows the format -MM. For example, the month of March would be inputted as 03.

Examples

storage_type$integer #"integer"

#> NULL